Bluetooth: Set device as temporary when link key is not stored.
When ever a device is created, device is set as temporary and device will be
set as permanent when a link key is generated which can be stored as per core
spec. The temporary device will be removed after ACL disconnection with the
remote device. At the same time, device needs to be set as temporary when one
more link key is generated which can not be stored. The changes take care of
the same.

Change-Id: I3b6e07208412507cc4e721fbf9dee0daa7546f33
CRs-fixed: 294318