Bluetooth: AVDTP: Support Flush Timeout for streaming connection
Support for flush timeout command is required for Streaming link
like A2DP. This is helpful in case of 2 or more ACL links. For
A2DP, if data is not sent within some time, it is better to
flush those packets, so that ACL credits could be used for other
ACL links.

Change-Id: I17e853b6eb1da470ea38c96f00cab371d70b5dd0
CRs-fixed: 325902