Bluez: Updating out_write timeout to higher value.
The out_write timeout is set for frames to be available after
write. This value is currently set to 1sec. If it exceeds 1 sec
there will be SUSPEND issued on the stream, which will be again
moved to start when write frames are available. Updating this value
to be in sync with AudioFlinger standby timeout. This avoids frequent
SUSPEND,START switching when RF performance is poor.

Change-Id: I1106434a5528102fdf1945ba1b3a37310eedbf71
CRs-Fixed: 330592