Bluez: Removed soft copy of caps data to setup list to avoid double free.
Found double free issue in failure cases for SET_CONFIGURATION. The reason
for this issue is using g_slist_copy API which copies the data ptr from source
to destination. This is ptr assignment and it leads issue when both the lists
try to do g_free on the data ptr. This issue is fixed by coping entire data
structure to destination ptr rather soft copy.

CRs-Fixed: 336743

Change-Id: Ieaa3cdbe8196a96acf91dac00bd0d89c0d91a0d7