Bluez: Proper error handling added in a2dp stream setup failure cases.
Found some crashes in a2dp in failure scenarios where
the remote device doesnt respond to start cmd etc., Found the root for
these crashes is callbacks registered from unix to a2dp are
not cleared when client free happens. When the callback registered is
not cleared, at later point of time the callback can hit for a client
that is already released. In such a case accessing some released pointer
will cause crash.

Change-Id: Id82e40ee6725b7bf9d1937febf18ece90925f414
CRs-Fixed: 336743