GATT: Make UUID compares (16 vs 128 bit) neutral.
For testability, it is handy to allow both UUID16 and UUID128s
to be used interchangably. UUID16 are BT abreviations of UUID128s
so no ambiguity is introduced by doing this, but for testing
purposes we would like to use UUID128s on occasion for certain
qualification tests. This affects find_by_type opcode only.

Change-Id: Ibad5ffa0c272967c8340bdffe7b98d255543f12e
CRs-Fixed: 359276