Bluetooth: Create stored device from lekeys
During BlueZ initialization, we need to check the lekeys database
to see if there is any LE device which needs to be created. Currently
we only look at primary service database to figure out which LE device
should be created which is not sufficient if we have not done a primary
search on a given LE device or if the primary search failed.

CRs-fixed: 368744
Change-Id: Icee4ae1b97738ec2af9967d0f9e4f3f5119e8f6a