Bluetooth: Avoid releasing local socket when there is connection collision.
Found when there is an incoming avrcp connection collision with outgoing
connection, both the sockets are being released. The incoming connection
will be updated with security_block in this case. But the outgoing connection
will not have any update and l2cap socket will be released locally. This
can cause reuse of the Channel ID info, which causes other l2cap connections
even going to bad state. The fix is to close outgoing connection only on
reply from remote device.

CRs-Fixed: 389663

Change-Id: I91fc8f57cd1536216b3a328a28a7deded0a97c65