Bluetooth: Setting authentication req to LOW for HID Mouse devices.
If a HID Mouse is 2.0 device, it doesn't support bonding proceedure.
If it is a 2.1 device, then authentication is mandatory. The current
change is to go ahead for dedicated bonding to a HID Mouse with
security level set to LOW. If the remote device features support
SSP pairing, the security level will be reset to HIGH and the
link key will be stored.

CRs-Fixed: 385463
Change-Id: I92eb4c5b3713de931a2d65d96b92d9412583fb9a