Bluetooth: Avoid AVRCP control based on A2dp Stream State.
AVRCP Pause is sent to Application layers when remote devices issues
Suspend command this is causing Video/Music App to Pause when there
is active call over BT and user tries to Play. Now we are Ignoring
sending the PAUSE to Application Layer when receive Suspend from
Remote device.

CRs-fixed: 396640

Change-Id: I66308bf1c299debfdf814f884656ea252abd57b5