Bluez: Increased delay in avctp connection request to avoid collision
Found that there is avctp connection collision which is initiated same
time from headset and DUT incase of incoming connection. To avoid this
collision increased the timer to delay the avctp connection request from
DUT to 6 seconds so that DUT can handle the conection request from the
headset successfully.

CRs-fixed: 443184
Change-Id: Ic0a865e3bca961830a2fb6b0e5da84eb9bae4d82